Overhaul tests, extract json_rpc logic to own module
This is a messy commit. Initially, the objective was just to extract the
JSON-RPC-related logic to its own module and add specific errors for each
JSON-RPC error code, but this revealed a number of underlying problems in the
structure of the module, which were fixed in the same process. Tests were
amended to cope with this change.